+++ packages/react/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @assistant-ui/react
 
+## 0.4.3
+
+### Minor Changes
+
+- feat: scrolling to bottom during streaming is now instant
+- fix: useSmooth gets triggered during branch switch
+
 ## 0.4.2
 
 ### Minor Changes